requirements.txt
requests==2.25.0
flask==1.1.1
urllib3==1.25.8
pyyaml==5.3.1
jinja2==2.11.2
cryptography==2.9

